+++ NEWS.md
@@ -1,3 +1,6 @@
+## gDRcore 1.3.14 - 2024-10-24
+* split and refactor annotation functions
+
 ## gDRcore 1.3.13 - 2024-10-21
 * add support for reprocessing data with old data model